variadic functions

+++ 0x10-variadic_functions/3-print_all.c
@@ -35,7 +35,7 @@ void print_all(const char * const format, ...)
 			printf("%d", va_arg(valist, int)), z = 1;
 			break;
 		case 'f':
-			printf("%f", va_arg(valist, double)), z = 1;
+			printf("%f", va_arg(valist, float)), z = 1;
 			break;
 		case 's':
 			string = va_arg(valist, char *), z = 1;
